pref: colorized output

+++ bakup.sh
@@ -6,6 +6,10 @@ FILES=(~/.vimrc)
 
 for file in ${FILES[@]}
 do
-  echo "copy $file to current folder"
-  cp "$file" .
+  if [ -f "$file" ];then
+    echo -e "\033[32mcopy $file to current folder \033[0m"
+    cp "$file" .
+  else
+    echo -e "\033[31m$file don't exist\033[0m"
+  fi
 done
